+++ docs/export.rst
@@ -6,14 +6,14 @@ Exporting and Importing Designs
 Exporting Hardware Designs
 --------------------------
 
-.. autofunction:: pyrtl.verilog.output_to_verilog
+.. autofunction:: pyrtl.inputoutput.output_to_verilog
 
 .. autofunction:: pyrtl.inputoutput.output_to_firrtl
 
 Exporting Testbenches
 ------------------------
 
-.. autofunction:: pyrtl.verilog.output_verilog_testbench
+.. autofunction:: pyrtl.inputoutput.output_verilog_testbench
 
 
 Importing Verilog
@@ -25,12 +25,12 @@ Importing Verilog
 Outputting for Visualization
 ---------------------------
 
-.. autofunction:: pyrtl.inputoutput.output_to_trivialgraph
-.. autofunction:: pyrtl.inputoutput.output_to_graphviz
-.. autofunction:: pyrtl.inputoutput.graphviz_detailed_namer
-.. autofunction:: pyrtl.inputoutput.output_to_svg
-.. autofunction:: pyrtl.inputoutput.block_to_graphviz_string
-.. autofunction:: pyrtl.inputoutput.block_to_svg
-.. autofunction:: pyrtl.inputoutput.trace_to_html
-.. autofunction:: pyrtl.inputoutput.net_graph
+.. autofunction:: pyrtl.visualization.output_to_trivialgraph
+.. autofunction:: pyrtl.visualization.output_to_graphviz
+.. autofunction:: pyrtl.visualization.graphviz_detailed_namer
+.. autofunction:: pyrtl.visualization.output_to_svg
+.. autofunction:: pyrtl.visualization.block_to_graphviz_string
+.. autofunction:: pyrtl.visualization.block_to_svg
+.. autofunction:: pyrtl.visualization.trace_to_html
+.. autofunction:: pyrtl.visualization.net_graph
