+++ pyrtl/simulation.py
@@ -12,7 +12,7 @@ from .core import working_block, PostSynthBlock, _PythonSanitizer
 from .wire import Input, Register, Const, Output, WireVector
 from .memory import RomBlock
 from .helperfuncs import check_rtl_assertions, _currently_in_jupyter_notebook
-from .verilog import _VerilogSanitizer
+from .inputoutput import _VerilogSanitizer
 
 # ----------------------------------------------------------------
 #    __                         ___    __
@@ -1144,7 +1144,7 @@ class SimulationTrace(object):
         """
         if _currently_in_jupyter_notebook():
             from IPython.display import display, HTML, Javascript  # pylint: disable=import-error
-            from .inputoutput import trace_to_html
+            from .visualization import trace_to_html
             htmlstring = trace_to_html(self, trace_list=trace_list, sortkey=_trace_sort_key)
             html_elem = HTML(htmlstring)
             display(html_elem)
